gcc warnings

+++ zmalloc.c
@@ -10,7 +10,7 @@ the GNU General Public License, version 2, 1991.
 ********************************************/
 
 /*
- * $MawkId: zmalloc.c,v 1.17 2010/07/24 11:49:50 tom Exp $
+ * $MawkId: zmalloc.c,v 1.18 2010/07/24 12:18:04 tom Exp $
  * @Log: zmalloc.c,v @
  * Revision 1.6  1995/06/06  00:18:35  mike
  * change mawk_exit(1) to mawk_exit(2)
@@ -169,12 +169,14 @@ record_ptr(PTR ptr, size_t size)
     item->ptr = ptr;
     item->size = size;
 
-    TRACE(("...record_ptr %p -> %p %lu\n", item, ptr, size));
+    TRACE(("...record_ptr %p -> %p %lu\n", item, ptr, (unsigned long) size));
     result = tsearch(item, &ptr_data, compare_ptr_data);
     assert(result != 0);
     assert(*result != 0);
 
-    TRACE(("->%p (%p %lu)\n", (*result), (*result)->ptr, (*result)->size));
+    TRACE(("->%p (%p %lu)\n",
+	   (*result), (*result)->ptr,
+	   (unsigned long) (*result)->size));
     ShowPtrData();
 }
 
@@ -187,13 +189,16 @@ finish_ptr(PTR ptr, size_t size)
     dummy.ptr = ptr;
     dummy.size = size;
 
-    TRACE(("finish_ptr %p -> %p %lu\n", &dummy, ptr, size));
+    TRACE(("finish_ptr %p -> %p %lu\n", &dummy, ptr, (unsigned long) size));
     item = tfind(&dummy, &ptr_data, compare_ptr_data);
 
     assert(item != 0);
     assert(*item != 0);
 
-    TRACE(("... %p -> %p %lu\n", (*item), (*item)->ptr, (*item)->size));
+    TRACE(("... %p -> %p %lu\n",
+	   (*item),
+	   (*item)->ptr,
+	   (unsigned long) (*item)->size));
 
     tdelete(item, &ptr_data, compare_ptr_data);
     ShowPtrData();
